+++ prueba/database/migrations/2024_04_17_211021_create_merchandisings_table.php
@@ -18,7 +18,7 @@ class CreateMerchandisingsTable extends Migration
             $table->text('image')->nullable();
             $table->integer('view_count');
             $table->bigInteger('stock');
-            $table->integer('type_product');
+            $table->unsignedBigInteger('type_product');
             $table->integer('merchandising_value');
             $table->timestamps();
